// product/src/src/app/model/orderItem.ts
export class OrderItem {
    id: number=0;
    orderId: number=0;
    productId: number=0;
    productName: string = ''
    productDescription: string = ''
    productImageFile: string = ''
    productUnitOfMeasure: string = ''
    quantity: number = 0;
    price: number = 0.0;
    status: string = ''
}